refactor(models): Split models.py into focused submodules
The 787-line models.py mixed dataset, config-spec, result, and summary
models. Split into a models/ package grouped by concern:

- sample.py: Sample
- specs.py: TargetSpec/GraderSpec/GateSpec/SuiteSpec families + gate helpers
- results.py: LettaMessageUnion, TargetResult, GradeResult, SampleResult,
  and the Usage/Timing/Error primitives (with their TimingStats/ErrorSummary
  aggregates kept co-located)
- summaries.py: PerRunSummary, ModelSummary, Summary, ModelRun, RunnerResult

models/__init__.py re-exports the full public surface (plus the
test-visible _compare/normalize_weights helpers), so all existing
`from letta_evals.models import X` call sites continue to work unchanged.
